db queries and routes - show events and one event, split out users into another file
